[implement/payline] Restructured payment processing

+++ apps/api/src/entities/payment/payment.controller.ts
@@ -1,17 +1,17 @@
 import { CRUDController } from '../../types';
 import { NextFunction, Request, Response } from 'express';
 import { PrismaClient } from '@prisma/client';
-import createMollieClient from '@mollie/api-client';
-import configuration from '../../../configuration';
 import { AWSsendEmail } from '../../tools/notifications.tools';
 import { generateConfirmationEmail } from '../../templates/email';
-
-const mollieClient = createMollieClient({ apiKey: configuration.MOLLIE_API_KEY });
+import MollieService from '../../tools/payment/mollie.service';
+import PaylineService from '../../tools/payment/payline.service';
+import { PaymentService } from '../../tools/payment/payment.service';
 
 const prisma = new PrismaClient();
-export const PaymentType = {
-    SUBSCRIPTION: 'SUBSCRIPTION',
-    ONETIME: 'ONETIME',
+
+const services = {
+    mollie: MollieService,
+    payline: PaylineService,
 };
 
 export default class PaymentController implements CRUDController {
@@ -48,8 +48,12 @@ export default class PaymentController implements CRUDController {
 
     static async getByPaymentId(req: Request, res: Response, next: NextFunction) {
         try {
-            const { paymentId } = req.params;
-            const payment = await mollieClient.payments.get(paymentId);
+            const { paymentId, provider } = req.params;
+
+            const service: PaymentService = services[provider];
+
+            const payment = await service.getPayment(paymentId);
+
             res.status(200).json(payment);
         } catch (error) {
             next(error);
@@ -58,122 +62,125 @@ export default class PaymentController implements CRUDController {
 
     static async create(req: Request, res: Response, next: NextFunction) {
         try {
-            const { subscriptionId, offerId } = req.body;
-            const { user } = req;
+            // const { subscriptionId, offerId } = req.body;
+            // const { user } = req;
+
+            res.sendStatus(501);
 
             // Request gives us: subscriptionId, offerId, user (session info)
             // We check that the subscriptionId is valid
-            const subscription = await prisma.subscription.findFirst({
-                where: {
-                    id: +subscriptionId,
-                },
-            });
-
-            // We check that the offerId is valid
-            // Todo: verify if offer is valid
-            const offer = await prisma.offer.findFirst({
-                where: {
-                    id: +offerId,
-                },
-            });
+            // const subscription = await prisma.subscription.findFirst({
+            //     where: {
+            //         id: +subscriptionId,
+            //     },
+            // });
+            //
+            // // We check that the offerId is valid
+            // // Todo: verify if offer is valid
+            // const offer = await prisma.offer.findFirst({
+            //     where: {
+            //         id: +offerId,
+            //     },
+            // });
 
             // we apply the offer to the subscription price
-            let { price } = subscription;
-            if (offer) {
-                //TODO verify that the offer is valid
-
-                if (offer.offerType === 'POURCENTAGE') {
-                    price = price - (price * offer.value) / 100;
-                } else {
-                    price = price - offer.value;
-                }
-            }
+            // let { price } = subscription;
+            // if (offer) {
+            //     //TODO verify that the offer is valid
+            //
+            //     if (offer.offerType === 'POURCENTAGE') {
+            //         price = price - (price * offer.value) / 100;
+            //     } else {
+            //         price = price - offer.value;
+            //     }
+            // }
 
             // we create a mollie customer
-            const customer = await mollieClient.customers.create({
-                name: user.firstName,
-                email: user.email,
-            });
+            // const customer = await mollieClient.customers.create({
+            //     name: user.firstName,
+            //     email: user.email,
+            // });
 
             let payment;
             // depending on the subscription type, we create a payment or a subscription
-            if (subscription.subscriptionType === PaymentType.ONETIME) {
-                // we create a payment
-                payment = await mollieClient.payments.create({
-                    amount: {
-                        value: price.toString(),
-                        currency: 'EUR',
-                    },
-                    description: 'My first API payment',
-                    redirectUrl: 'https://yourwebshop.example.org/order/123456',
-                    webhookUrl: 'https://yourwebshop.example.org/webhook',
-                    metadata: {
-                        subscriptionId,
-                        offerId,
-                    },
-                    customerId: customer.id,
-                });
-            } else {
-                // we create a subscription
-                payment = await mollieClient.customers_subscriptions.create({
-                    amount: {
-                        value: price.toString(),
-                        currency: 'EUR',
-                    },
-                    customerId: customer.id,
-                    webhookUrl: `${process.env.APP_URL}/api/payments/webhook`,
-                    testmode: process.env.NODE_ENV === 'development',
-                    description: subscription.description,
-                    interval: subscription.refreshRate.toString(), // invalid
-                    startDate: new Date().toISOString(),
-                });
-            }
+            // if (subscription.subscriptionType === PaymentType.ONETIME) {
+            //     // we create a payment
+            //     // payment = await mollieClient.payments.create({
+            //     //     amount: {
+            //     //         value: price.toString(),
+            //     //         currency: 'EUR',
+            //     //     },
+            //     //     description: 'My first API payment',
+            //     //     redirectUrl: 'https://yourwebshop.example.org/order/123456',
+            //     //     webhookUrl: 'https://yourwebshop.example.org/webhook',
+            //     //     metadata: {
+            //     //         subscriptionId,
+            //     //         offerId,
+            //     //     },
+            //     //     customerId: customer.id,
+            //     // });
+            // } else {
+            //     // we create a subscription
+            //     // payment = await mollieClient.customers_subscriptions.create({
+            //     //     amount: {
+            //     //         value: price.toString(),
+            //     //         currency: 'EUR',
+            //     //     },
+            //     //     customerId: customer.id,
+            //     //     webhookUrl: `${process.env.APP_URL}/api/payments/webhook`,
+            //     //     testmode: process.env.NODE_ENV === 'development',
+            //     //     description: subscription.description,
+            //     //     interval: subscription.refreshRate.toString(), // invalid
+            //     //     startDate: new Date().toISOString(),
+            //     // });
+            // }
 
             // we store the payment id and the customer id in the database
             // Todo: does the user already have a subscription?
             //  if so, we need to update the subscription
-            await prisma.userSubscription.create({
-                data: {
-                    User: {
-                        connect: {
-                            id: user.id,
-                        },
-                    },
-                    Subscription: {
-                        connect: {
-                            id: subscription.id,
-                        },
-                    },
-                    price: price,
-                    paymentId: payment.id,
-                    customerId: customer.id,
-                    status: 'PENDING',
-                },
-            });
+            // await prisma.userSubscription.create({
+            //     data: {
+            //         User: {
+            //             connect: {
+            //                 id: user.id,
+            //             },
+            //         },
+            //         Subscription: {
+            //             connect: {
+            //                 id: subscription.id,
+            //             },
+            //         },
+            //         price: price,
+            //         paymentId: payment.id,
+            //         customerId: customer.id,
+            //         status: 'PENDING',
+            //     },
+            // });
 
-            await prisma.offerUsage.create({
-                data: {
-                    User: {
-                        connect: {
-                            id: user.id,
-                        },
-                    },
-                    Offer: {
-                        connect: {
-                            id: offer.id,
-                        },
-                    },
-                },
-            });
-
-            res.status(200).json({
-                paymentUrl: payment.getPaymentUrl(),
-            });
+            // await prisma.offerUsage.create({
+            //     data: {
+            //         User: {
+            //             connect: {
+            //                 id: user.id,
+            //             },
+            //         },
+            //         Offer: {
+            //             connect: {
+            //                 id: offer.id,
+            //             },
+            //         },
+            //     },
+            // });
+            //
+            // res.status(200).json({
+            //     paymentUrl: payment.getPaymentUrl(),
+            // });
         } catch (error) {
             next(error);
         }
     }
 
+    // @deprecated
     static async update(req: Request, res: Response, next: NextFunction) {
         try {
             const { id } = req.params;
@@ -195,7 +202,7 @@ export default class PaymentController implements CRUDController {
 
     static async delete(req: Request, res: Response, next: NextFunction) {
         try {
-            const { id } = req.params;
+            const { id, provider } = req.params;
             const { user } = req;
             const payment = await prisma.userSubscription.findFirst({
                 where: {
@@ -207,9 +214,13 @@ export default class PaymentController implements CRUDController {
             if (!payment) {
                 throw new Error('Payment not found');
             }
-            await mollieClient.customers_subscriptions.cancel(payment.paymentId, {
-                customerId: payment.customerId,
-            });
+            const service: PaymentService = services[provider];
+
+            if (!service) {
+                throw new Error('Provider not found');
+            }
+
+            await service.cancelSubscription(payment.customerId, payment.paymentId);
             await prisma.userSubscription.delete({
                 where: {
                     id: +id,
@@ -221,10 +232,10 @@ export default class PaymentController implements CRUDController {
         }
     }
 
-    static async webhooksStatus(req: Request, res: Response) {
+    static async mollieWebhooksStatus(req: Request, res: Response) {
         const { id } = req.body;
 
-        const status = await mollieClient.payments.get(id);
+        const status = await MollieService.getPayment(id);
 
         if (!status) {
             return res.sendStatus(200);
@@ -272,5 +283,63 @@ export default class PaymentController implements CRUDController {
                 status: status.status,
             },
         });
+    }
+
+    //createRefund
+    static async createRefund(req: Request, res: Response, next: NextFunction) {
+        try {
+            const { id, provider } = req.params;
+            const { user } = req;
+            const payment = await prisma.userSubscription.findFirst({
+                where: {
+                    id: +id,
+                    userId: +user.id,
+                    status: 'ACTIVE',
+                },
+            });
+            if (!payment) {
+                throw new Error('Payment not found');
+            }
+            const service: PaymentService = services[provider];
+            if (!service) {
+                throw new Error('Provider not found');
+            }
+
+            const refund = await service.createRefund(payment.customerId, payment.paymentId);
+            res.status(200).json(refund);
+        } catch (error) {
+            next(error);
+        }
+    }
+
+    //paylineWebhooksStatus
+    static async paylineWebhooksStatus(req: Request, res: Response) {
+        // const { id } = req.body;
+
+        // const status = await PaylineService.getPayment(id);
+        //
+        // if (!status) {
+        //     return res.sendStatus(200);
+        // }
+        //
+        // const subscription = await prisma.userSubscription.findFirst({
+        //     where: {
+        //         paymentId: id,
+        //     },
+        //     select: {
+        //         Subscription: true,
+        //         User: true,
+        //         id: true,
+        //         createdAt: true,
+        //         updatedAt: true,
+        //         price: true,
+        //     },
+        // });
+        //
+        // if (!subscription) {
+        //     return res.sendStatus(200);
+        // }
+
+        res.sendStatus(501);
     }
 }
